+++ packages/core-quotations/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'unchained:core-quotations',
-  version: '0.53.2',
+  version: '0.53.3',
   summary: 'Unchained Engine Core: Quotations',
   git: 'https://github.com/unchainedshop/unchained',
   documentation: 'README.md',
@@ -17,14 +17,14 @@ Package.onUse((api) => {
   api.use('dburles:collection-helpers@1.1.0');
   api.use('aldeed:collection2@3.2.1');
 
-  api.use('unchained:core-files@0.53.2');
-  api.use('unchained:core@0.53.2');
-  api.use('unchained:utils@0.53.2');
-  api.use('unchained:core-users@0.53.2');
-  api.use('unchained:core-products@0.53.2');
-  api.use('unchained:core-countries@0.53.2');
-  api.use('unchained:core-logger@0.53.2');
-  api.use('unchained:core-worker@0.53.2');
+  api.use('unchained:core-files@0.53.3');
+  api.use('unchained:core@0.53.3');
+  api.use('unchained:utils@0.53.3');
+  api.use('unchained:core-users@0.53.3');
+  api.use('unchained:core-products@0.53.3');
+  api.use('unchained:core-countries@0.53.3');
+  api.use('unchained:core-logger@0.53.3');
+  api.use('unchained:core-worker@0.53.3');
 
   api.mainModule('quotations.js', 'server');
 });