+++ packages/roles/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   name: 'unchained:roles',
-  version: '0.53.2',
+  version: '0.53.3',
   summary: 'Unchained Engine: Roles',
   git: 'https://github.com/unchainedshop/unchained',
   documentation: 'README.md',